HWKSC-1320: feat: Add Machine Support Tickets page
This commit adds a new file `tickets.jsx` under the `src/page/product/support` directory. The file contains a functional component `MachineSupportTickets` that renders the `MachineLayout` component with the `TicketsTab` component inside. This addition introduces a new page for managing machine support tickets in the product section of the application.