feat: Update customerDefaultValues function to include contacts parameter
This commit updates the `customerDefaultValues` function in the `default-values.jsx` file. The changes include adding a new `contacts` parameter to the function signature and using it to set the `contacts` property in the returned default values object. This modification allows for more flexibility in providing custom contacts data when calling the `customerDefaultValues` function.