refactor: Add TICKET icon to ICON_WEB
This commit adds the TICKET icon to the ICON_WEB object in the icon-directory.js file. The TICKET icon will be used to represent tickets in the web application. This change improves the consistency and completeness of the available icons in the application.

Note: This commit message follows the established convention of using a prefix (`refactor`) to indicate code refactoring or improvements.