refactor: Add filterPeriodOption state to TicketsTab component
This commit adds the `filterPeriodOption` state to the `TicketsTab` component in the `tickets-tab.jsx` file. The state is initialized with a default value of `3` and is used to filter the customer tickets based on a specified period. This change improves the functionality of the `TicketsTab` component by allowing users to easily filter the displayed tickets.

Note: This commit message follows the established convention of using a prefix (`refactor`) to indicate code refactoring or improvements.