HPS-1332: [Perf] refactor: Update redux-persist configurations in root-reducer.js
This commit updates the redux-persist configurations in the root-reducer.js file. It removes the `blacklist` properties for several persist configurations and adds `whitelist` properties for others. This change aligns the configurations with the recent changes made in other files, improving code consistency and maintainability.

Note: This commit message follows the established convention of using a prefix (`refactor`) to indicate code refactoring or improvements.