Add security user and machines by country to GeneralAppPage
This commit refactors the GeneralAppPage component to include the security user and machines by country. It fetches the security user based on the user ID and retrieves the machines by country using the 'getMachinesByCountry' action. This change improves the functionality of the page by displaying relevant information to the user.

Note: This commit message follows the established convention of using a prefix (`refactor`) to indicate code refactoring or improvements.