HPS-1224: Add websocket event constants for Housekeeping.
This commit adds the websocket event constants to the project. The new file 'websocket-event.js' contains a list of websocket events that can be used throughout the application. This addition enhances the functionality of the project by providing a centralized location for managing websocket events.

Note: This commit message follows the established convention of using a prefix (`feat`) to indicate a new feature or enhancement.